// src/server/types/WishlistProduct.type.ts
export type GetProductVariantsForWishlistSchema = {
    ProductId: string;
    ProductVariantId: string;
    WishlistProductId: string;
    ProductName: string;
    ProductLabel: string;
    ProductCategoryName: string;
    ProductCategoryLabel: string;
    ProductVariantName: string;
    OriginalPrice: number;
    Quantity: number;
    Discount: number;
    LatestPrice: number;
    VariantCategoryName: string;
    VariantCategoryLabel: string;
    VariantCategoryValueName: string;
    VariantCategoryValueLabel: string;
    ProductImageId: string;
}